// src/generated/models/Cgmp_notificationrulesModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */
export const Cgmp_notificationrulesstatecode = {
  0: 'Active',
  1: 'Inactive'
} as const;
export type Cgmp_notificationrulesstatecode = keyof typeof Cgmp_notificationrulesstatecode;
export const Cgmp_notificationrulesstatuscode = {
  1: 'Active',
  2: 'Inactive'
} as const;
export type Cgmp_notificationrulesstatuscode = keyof typeof Cgmp_notificationrulesstatuscode;

export interface Cgmp_notificationrulesBase {
  cgmp_conditions?: string;
  cgmp_eventtype?: string;
  cgmp_isactive?: boolean;
  cgmp_name?: string;
  cgmp_notificationruleid: string;
  cgmp_userid: string;
  importsequencenumber?: number;
  overriddencreatedon?: string;
  ownerid: string;
  owneridtype: string;
  statecode: Cgmp_notificationrulesstatecode;
  statuscode?: Cgmp_notificationrulesstatuscode;
  timezoneruleversionnumber?: number;
  utcconversiontimezonecode?: number;
}

export interface Cgmp_notificationrules extends Cgmp_notificationrulesBase {
  cgmp_isactivename?: string;
  createdbyname?: string;
  createdbyyominame: string;
  createdon?: string;
  createdonbehalfbyname?: string;
  createdonbehalfbyyominame: string;
  modifiedbyname?: string;
  modifiedbyyominame: string;
  modifiedon?: string;
  modifiedonbehalfbyname?: string;
  modifiedonbehalfbyyominame: string;
  owneridname: string;
  owneridyominame: string;
  owningbusinessunitname: string;
  statecodename?: string;
  statuscodename?: string;
  versionnumber?: number;
  createdby?: object;
  _createdby_value?: string;
  createdonbehalfby?: object;
  _createdonbehalfby_value?: string;
  modifiedby?: object;
  _modifiedby_value?: string;
  modifiedonbehalfby?: object;
  _modifiedonbehalfby_value?: string;
  owningbusinessunit?: object;
  _owningbusinessunit_value?: string;
  owningteam?: object;
  _owningteam_value?: string;
  owninguser?: object;
  _owninguser_value?: string;
}